+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.0-rc.18",
-  genVersion: "2.723.8",
-  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.18 2.723.8 2.0 @orq-ai/node",
+  sdkVersion: "3.14.0-rc.19",
+  genVersion: "2.723.11",
+  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.19 2.723.11 2.0 @orq-ai/node",
 } as const;